pyproject.toml
[project]
name = "ml-inference"
version = "0.1.0"
description = "ML inference benchmark: PyTorch LSTM via FastAPI and gRPC"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "numpy",
    "fastapi",
    "uvicorn[standard]",
    "psutil",
    "httpx",
    "grpcio",
    "grpcio-tools",
    "pyyaml",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
